pyproject.toml
[project]
name = "camping-bot"
version = "0.1.0"
description = "Multi-site camping reservation bot framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "playwright>=1.51.0",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.28.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
